refactor: split minimap panel chrome and overture overlay build (#3878)
Extract minimap visible-panel/gesture bus emitters and overture overlay
build/submit chrome into library parts; document layout in repo spec.

Refs #3878
